+++ beets/ui/__init__.py
@@ -110,6 +110,27 @@ def print_(*strings):
         txt = txt.encode(_encoding(), 'replace')
     print(txt)
 
+def input_(prompt=None):
+    """Like `raw_input`, but decodes the result to a Unicode string.
+    Raises a UserError if stdin is not available. The prompt is sent to
+    stdout rather than stderr. A printed between the prompt and the
+    input cursor.
+    """
+    # raw_input incorrectly sends prompts to stderr, not stdout, so we
+    # use print() explicitly to display prompts.
+    # http://bugs.python.org/issue1927
+    if prompt:
+        if isinstance(prompt, unicode):
+            prompt = prompt.encode(_encoding(), 'replace')
+        print(prompt, end=' ')
+
+    try:
+        resp = raw_input()
+    except EOFError:
+        raise UserError('stdin stream ended while input required')
+
+    return resp.decode(sys.stdin.encoding, 'ignore')
+
 def input_options(options, require=False, prompt=None, fallback_prompt=None,
                   numrange=None, default=None, color=False, max_width=72):
     """Prompts a user for input. The sequence of `options` defines the
@@ -241,11 +262,7 @@ def input_options(options, require=False, prompt=None, fallback_prompt=None,
             fallback_prompt += '%i-%i, ' % numrange
         fallback_prompt += ', '.join(display_letters) + ':'
 
-    # raw_input incorrectly sends prompts to stderr, not stdout, so we
-    # use print() explicitly to display prompts.
-    # http://bugs.python.org/issue1927
-    print(prompt, end=' ')
-    resp = raw_input()
+    resp = input_(prompt)
     while True:
         resp = resp.strip().lower()
 
@@ -273,8 +290,7 @@ def input_options(options, require=False, prompt=None, fallback_prompt=None,
                 return resp
 
         # Prompt for new input.
-        print(fallback_prompt, end=' ')
-        resp = raw_input()
+        resp = input_(fallback_prompt)
 
 def input_yn(prompt, require=False, color=False):
     """Prompts the user for a "yes" or "no" response. The default is